move argparse into utils
